yshop2.0.2,修复issue#I1B8RB,#I1B49Y,#I1BASM,#I1B9O7,#I1B9ED

+++ yshop-shop/src/main/java/co/yixiang/modules/shop/service/impl/YxStoreCategoryServiceImpl.java
@@ -1,8 +1,11 @@
 package co.yixiang.modules.shop.service.impl;
 
 import cn.hutool.core.util.ObjectUtil;
+import co.yixiang.exception.BadRequestException;
 import co.yixiang.modules.shop.domain.YxStoreCategory;
+import co.yixiang.modules.shop.domain.YxStoreProduct;
 import co.yixiang.modules.shop.repository.YxStoreCategoryRepository;
+import co.yixiang.modules.shop.repository.YxStoreProductRepository;
 import co.yixiang.modules.shop.service.YxStoreCategoryService;
 import co.yixiang.modules.shop.service.dto.YxStoreCategoryDTO;
 import co.yixiang.modules.shop.service.dto.YxStoreCategoryQueryCriteria;
@@ -31,13 +34,16 @@ import java.util.stream.Collectors;
 @Transactional(propagation = Propagation.SUPPORTS, readOnly = true, rollbackFor = Exception.class)
 public class YxStoreCategoryServiceImpl implements YxStoreCategoryService {
 
-
     private final YxStoreCategoryRepository yxStoreCategoryRepository;
+    private final YxStoreProductRepository yxStoreProductRepository;
 
     private final YxStoreCategoryMapper yxStoreCategoryMapper;
 
-    public YxStoreCategoryServiceImpl(YxStoreCategoryRepository yxStoreCategoryRepository, YxStoreCategoryMapper yxStoreCategoryMapper) {
+    public YxStoreCategoryServiceImpl(YxStoreCategoryRepository yxStoreCategoryRepository,
+                                      YxStoreProductRepository yxStoreProductRepository,
+                                      YxStoreCategoryMapper yxStoreCategoryMapper) {
         this.yxStoreCategoryRepository = yxStoreCategoryRepository;
+        this.yxStoreProductRepository = yxStoreProductRepository;
         this.yxStoreCategoryMapper = yxStoreCategoryMapper;
     }
 
@@ -92,6 +98,12 @@ public class YxStoreCategoryServiceImpl implements YxStoreCategoryService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void delete(Integer id) {
+        YxStoreCategory storeCategory = yxStoreCategoryRepository.findByPid(id);
+        if(storeCategory != null) throw new BadRequestException("请先删除子类");
+
+        List<YxStoreProduct> storeProduct = yxStoreProductRepository.findByCateId(String.valueOf(id));
+        if(!storeProduct.isEmpty()) throw new BadRequestException("此分类下有商品,不能删除");
+
         yxStoreCategoryRepository.deleteById(id);
     }
